Refactor `create_sequential_variables()`
I've been fiddling around with this study a fair bit as part of
investigating the performance issues and it occured to me that there
might be a simpler abstraction here than the one I originally had.

I'm not particularly expecting you to merge this (so feel free to close)
but I thought it might be useful for inspiration if nothing else.